@@ -5,6 +5,7 @@
  */
 package ifpe.edu.br.servsimples.ui.services;
 
+import android.content.Context;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
@@ -17,6 +18,7 @@ import android.widget.ListView;
 import android.widget.Spinner;
 import android.widget.Toast;
 
+import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 
 import java.util.ArrayList;
@@ -28,41 +30,41 @@ import ifpe.edu.br.servsimples.managers.IServerManagerInterfaceWrapper;
 import ifpe.edu.br.servsimples.managers.ServerManager;
 import ifpe.edu.br.servsimples.model.Service;
 import ifpe.edu.br.servsimples.model.User;
+import ifpe.edu.br.servsimples.ui.UIInterfaceWrapper;
 import ifpe.edu.br.servsimples.util.PersistHelper;
+import ifpe.edu.br.servsimples.util.ServSimplesAppLogger;
 import ifpe.edu.br.servsimples.util.ServerResponseCodeParser;
 
 public class SearchServiceFragment extends Fragment {
 
     private static final String TAG = SearchServiceFragment.class.getSimpleName();
 
+    private UIInterfaceWrapper.FragmentUtil mFragmentUtil;
     private static final int GET_CATEGORIES_OK = 0;
     private static final int GET_CATEGORIES_FAIL = 1;
     private static final int GET_SERVICES_OK = 2;
+    private int mLastCategorySelected = 0;
     private CategoriesAdapter mCategoriesAdapter;
     private ServiceListViewAdapter mServiceAdapter;
     private ListView mServiceListView;
     private List<String> mCategories = new ArrayList<>(Collections.singletonList("Default"));
     private List<Service> mServices = new ArrayList<>();
-
     private Spinner mSpCategories;
 
     private final Handler mHandler = new Handler(Looper.getMainLooper()) {
         public void handleMessage(Message message) {
             final int what = message.what;
             switch (what) {
-                case GET_CATEGORIES_OK:
+                case GET_CATEGORIES_OK -> {
                     mCategoriesAdapter = new CategoriesAdapter(getContext(), mCategories);
                     mSpCategories.setAdapter(mCategoriesAdapter);
-                    break;
-
-                case GET_CATEGORIES_FAIL:
-                    requireActivity().finish();
-                    break;
-
-                case GET_SERVICES_OK:
+                    mSpCategories.setSelection(mLastCategorySelected);
+                }
+                case GET_CATEGORIES_FAIL -> requireActivity().finish();
+                case GET_SERVICES_OK -> {
                     mServiceAdapter = new ServiceListViewAdapter(getContext(), mServices);
                     mServiceListView.setAdapter(mServiceAdapter);
-                    break;
+                }
             }
         }
     };
@@ -79,6 +81,11 @@ public class SearchServiceFragment extends Fragment {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    public void onResume() {
+        super.onResume();
         retrieveCategoriesInfo();
     }
 
@@ -139,6 +146,17 @@ public class SearchServiceFragment extends Fragment {
 
     private void setListeners() {
         setServiceCategorySelectionListener();
+        setServiceListItemListener();
+    }
+
+    private void setServiceListItemListener() {
+        mServiceListView.setOnItemClickListener((parent, view, position, id) -> {
+            if (ServSimplesAppLogger.ISLOGABLE)
+                ServSimplesAppLogger.d(TAG, "onItemSelected: pos:" + position
+                        + " service name:" + mServices.get(position).getName());
+            Fragment serviceDetail = ServiceDetailFragment.newInstance(mServices.get(position));
+            mFragmentUtil.openFragment(serviceDetail, true);
+        });
     }
 
     private void setServiceCategorySelectionListener() {
@@ -146,6 +164,7 @@ public class SearchServiceFragment extends Fragment {
             @Override
             public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
                 retrieveServicesInfoForCategory(mCategories.get(position));
+                mLastCategorySelected = position;
             }
 
             @Override
@@ -158,5 +177,13 @@ public class SearchServiceFragment extends Fragment {
     private void findViewsById(View view) {
         mSpCategories = view.findViewById(R.id.sp_findservice);
         mServiceListView = view.findViewById(R.id.lv_findservice);
+    }
+
+    @Override
+    public void onAttach(@NonNull Context context) {
+        if (context instanceof UIInterfaceWrapper.FragmentUtil) {
+            mFragmentUtil = (UIInterfaceWrapper.FragmentUtil) context;
+        }
+        super.onAttach(context);
     }
 }
